Next diagrams data moved to config clas
