+++ src/main/webapp/src/app/diagrams/diagrams.component.html
@@ -8,7 +8,7 @@
                        #grid [showGrid]="showGrid" (gridFull)="onGridFull($event)"
                        (widgetPositionChange)="onWidgetChange($event)"
                        style="height:93%;width:97%;margin-top:-15px">
-        <ngx-widget *ngFor="let widget of this.initSetup.widgets;let $index = index;" [position]="widget" [movable]="editable"
+        <ngx-widget *ngFor="let widget of initSetup.widgets;let $index = index;" [position]="widget" [movable]="editable"
                     [resizable]="editable">
           <div style="height:100%;width:100%; display:flex;">
             <div class="search-result">
@@ -111,28 +111,28 @@
             </div>
           </div>
         </ngx-widget>
-        <ngx-widget *ngFor="let widget of this.initSetup.widgets2;let $index = index;" [position]="widget" [movable]="editable"
+        <ngx-widget *ngFor="let widget of initSetup.widgets2;let $index = index;" [position]="widget" [movable]="editable"
                     [resizable]="editable">
           <div style="height:100%;width:100%; display:flex;">
-            <nvd3 style="fill:white;margin: 5 auto;" [options]="options2" [data]="data2"></nvd3>
+            <nvd3 style="fill:white;margin: 5 auto;" [options]="initSetup.options2" [data]="initSetup.data2"></nvd3>
           </div>
         </ngx-widget>
-        <ngx-widget *ngFor="let widget of this.initSetup.widgets3;let $index = index;" [position]="widget" [movable]="editable"
+        <ngx-widget *ngFor="let widget of initSetup.widgets3;let $index = index;" [position]="widget" [movable]="editable"
                     [resizable]="editable">
           <div style="height:100%;width:100%; display:flex;">
-            <nvd3 style="fill:white;" [options]="options3" [data]="data3"></nvd3>
+            <nvd3 style="fill:white;" [options]="initSetup.options3" [data]="initSetup.data3"></nvd3>
           </div>
         </ngx-widget>
-        <ngx-widget *ngFor="let widget of this.initSetup.widgets4;let $index = index;" [position]="widget" [movable]="editable"
+        <ngx-widget *ngFor="let widget of initSetup.widgets4;let $index = index;" [position]="widget" [movable]="editable"
                     [resizable]="editable">
           <div style="height:100%;width:100%; display:flex;">
-            <nvd3 style="fill:white;" [options]="options3" [data]="data3"></nvd3>
+            <nvd3 style="fill:white;" [options]="initSetup.options3" [data]="initSetup.data3"></nvd3>
           </div>
         </ngx-widget>
-        <ngx-widget *ngFor="let widget of this.initSetup.widgets5;let $index = index;" [position]="widget" [movable]="editable"
+        <ngx-widget *ngFor="let widget of initSetup.widgets5;let $index = index;" [position]="widget" [movable]="editable"
                     [resizable]="editable">
           <div style="height:100%;width:100%; display:flex;">
-            <nvd3 style="fill:white;" [options]="options5" [data]="data5"></nvd3>
+            <nvd3 style="fill:white;" [options]="initSetup.options5" [data]="initSetup.data5"></nvd3>
           </div>
         </ngx-widget>
       </ngx-widget-grid>
